# perf(pnp): optimize `findPackageLocator`