**What's the problem this PR addresses?**

`findPackageLocator` is used by every function in the `pnpapi` and could be faster

**How did you fix it?**

Removed the `packageLocationLengths` logic and replaced it with `lastIndexOf` to check the sections of the filepath

**Result**

Time spent in `findPackageLocator`
```diff
# yarn build:doc
- 36472 calls, 541.3566000048304 ms
+ 36294 calls, 472.3983999992197 ms

# Webpack build from a private project with enableGlobalCache: true
- 27247 calls, 631.8915001358837 ms
+ 27255 calls, 439.93869973160326 ms
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.